Removed debugging statements.  Made handle_gl_errors not print anything on release mode.
Former-commit-id: 6fde82c58ca681f4ae0c36fa73138406881b989c [formerly dc5341b32d1b1ad80ff8f85be08fe7188802f606] [formerly bc980c8d21fc156e32637db3846500cf2a1dcf9d] [formerly 1136ff0a712bbe7d231f26f5a3da03e80ca0b14b] [formerly 94301adf1b73aa27b8c925fc0e75a2a6ec12031f [formerly f0f512456a2870592ccb57f0c8c56e9a038bc822 [formerly a56f8aa6836e1319f330aef3cce99392f7353c93] [formerly 94301adf1b73aa27b8c925fc0e75a2a6ec12031f [formerly 94301adf1b73aa27b8c925fc0e75a2a6ec12031f [formerly 94301adf1b73aa27b8c925fc0e75a2a6ec12031f [formerly 94301adf1b73aa27b8c925fc0e75a2a6ec12031f [formerly 94301adf1b73aa27b8c925fc0e75a2a6ec12031f [formerly c339e4582fbfe422a0f919166f112e1eb84ad1fd]]]]]]]] [formerly 87b0d5961fc79b7278a8e6d07b54fe7a23a70142 [formerly 2cbe183d2fc5fd39268ddb258e07873839090eb8] [formerly 5f562a524b6463bee091ad65cdee8716fea1ad92] [formerly f55767552f7faae79bcec9af1bc8fa0ec8774edf] [formerly c5cc2fcea314487d9608dcca738f4149b8228eba] [formerly ccf9f55f26328665cc50d1c6a37bf8fc3feb5014] [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a] [formerly 77162e9c975b68c76ce82d8b1eac80444a7c0979 [formerly ffb0ec9f3e3f33a43a19c013cfc388843ca580e2] [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ab0eeb44febf61f3f8f7f1717bcd31c6edc8500e]]]]]]] [formerly 4ed621baf65556804f8ba2c251c4a115c2dae31b [formerly 4ed621baf65556804f8ba2c251c4a115c2dae31b [formerly 4ed621baf65556804f8ba2c251c4a115c2dae31b [formerly 4ed621baf65556804f8ba2c251c4a115c2dae31b [formerly 4ed621baf65556804f8ba2c251c4a115c2dae31b [formerly 4ed621baf65556804f8ba2c251c4a115c2dae31b [formerly eeb3a5b0ee3cf157dca9de00f40642b927272e12]]]]]]]]
Former-commit-id: 87b0d5961fc79b7278a8e6d07b54fe7a23a70142 [formerly 2cbe183d2fc5fd39268ddb258e07873839090eb8] [formerly 5f562a524b6463bee091ad65cdee8716fea1ad92] [formerly f55767552f7faae79bcec9af1bc8fa0ec8774edf] [formerly c5cc2fcea314487d9608dcca738f4149b8228eba] [formerly ccf9f55f26328665cc50d1c6a37bf8fc3feb5014] [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a] [formerly 77162e9c975b68c76ce82d8b1eac80444a7c0979 [formerly ffb0ec9f3e3f33a43a19c013cfc388843ca580e2] [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ec14e62ba43ce7ecb7528f39b5d5fa42c294662a [formerly ab0eeb44febf61f3f8f7f1717bcd31c6edc8500e]]]]]]]
Former-commit-id: f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly 8fd9095b0695793e05ecc715f31bef8965ebcf2a]]]] [formerly 12251d794fe76bbe2632683540c56be612f6dcdb [formerly 51163fad2b267df9ad0df68cc4e4d93867dfce27] [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly 8fd9095b0695793e05ecc715f31bef8965ebcf2a]]]] [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly 8fd9095b0695793e05ecc715f31bef8965ebcf2a]]] [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly 8fd9095b0695793e05ecc715f31bef8965ebcf2a]] [formerly f4d0ef048b95f3bdf8f332c697c1ad3fac6315fd [formerly 8fd9095b0695793e05ecc715f31bef8965ebcf2a] [formerly 8fd9095b0695793e05ecc715f31bef8965ebcf2a [formerly f9e6e96cff9ea594c9492d9e7ba26f97b733c31a]]]]]]]
Former-commit-id: bb719e2464ff9b456e0bca5b4c9d055faec92bc3
Former-commit-id: 6980f3ba1983139e456d4aa23e8bd3c34a1d7c79
Former-commit-id: e988e7b884ca878463185ba9d5041a1d30df2784
Former-commit-id: c05d59a37e36ebe390741d2772b66d6adde02078
Former-commit-id: 62902e37ae331472fc85b26d1652f37732310196 [formerly bfc8702adfd4b1723e48150c2b07c53e0054a169] [formerly 6f06f34e5f6357b42b4f3faca1918c3643c1b557 [formerly 346b6ff9cb7a5b5ac854731db976c4c28a0b3f6a]]
Former-commit-id: a80977b4c6aeb92faf7f39f0030cfb4a93178422 [formerly 4d956148e5b49c62a6ffa9e95b7b64bed0074582]
Former-commit-id: 26031908cceb5a3c8cb1d77d8e9ad5c61421ee17

+++ mrViewer/src/video/mrvGLQuad.cpp
@@ -1006,19 +1006,19 @@ namespace mrv {
 	_shader->setTextureUnit( "lut", 3 );
 	CHECK_GL;
 
-        _shader->setUniform( "mask", (int) _mask );
+        _shader->setUniform( "mask", _mask );
 	CHECK_GL;
-        _shader->setUniform( "mask_value", (int) _mask_value );
+        _shader->setUniform( "mask_value", _mask_value );
 	CHECK_GL;
          // if ( _view->stereo_input() & CMedia::kLeftRightStereoInput )
          //     _shader->setUniform( "width", (int) (_width/2) );
          // else
-             _shader->setUniform( "width", (int) _width );
+        _shader->setUniform( "width", _width );
 	CHECK_GL;
          // if ( _view->stereo_input() & CMedia::kTopBottomStereoInput )
          //     _shader->setUniform( "height", (int) (_height/2) );
          // else
-             _shader->setUniform( "height", (int) _height );
+        _shader->setUniform( "height", _height );
 	CHECK_GL;
 
 	_shader->setUniform( "gain",  _view->gain() );
